[tech] Extracted viewer from photos for use in drive
