+++ src/components/Image/index.js
@@ -1,0 +1,1 @@
+export { default as ImageLoader } from './ImageLoader'
