+++ src/viewer/index.js
@@ -1,0 +1,1 @@
+export { default } from './Viewer'
