+++ src/viewer/viewer.styl
@@ -37,11 +37,11 @@
         pointer-events none
 
 .pho-viewer-nav-next
-    background-image  embedurl('../assets/icons/icon-arrow-right.svg')
+    background-image  embedurl('./assets/icon-arrow-right.svg')
     right             0
 
 .pho-viewer-nav-previous
-    background-image  embedurl('../assets/icons/icon-arrow-left.svg')
+    background-image  embedurl('./assets/icon-arrow-left.svg')
     left              0
 
 .pho-viewer-photo
